Add files via upload
This is provided in .py, however you will need to edit the text so file in the first line is set to your file directory. It only works with .edf format. You can also edit how much of the EEG is read by manipulating the code. With the comments, it should be easy to find. If you need EEGs to run I recommend this free academic resource: https://www.isip.piconepress.com/projects/tuh_eeg/
I plan on adding more to the script. This was done as a grad school project so it is relatively barebones. Please message with any questions or concerns.